show gui maxenemycount

+++ ButiEngine_User/WaveManager.h
@@ -30,8 +30,8 @@ namespace ButiEngine {
 		void AddSpawnCount();
 
 		bool IsClearAnimationFlag() { return m_isLastWaveClear; }
-		std::int16_t GetSpawnCount() { return m_enemySpawnCount; }
-		std::int16_t GetMaxEnemyCount() { return m_maxEnemyCount; }
+		std::int32_t GetSpawnCount() { return m_enemySpawnCount; }
+		std::int32_t GetMaxEnemyCount() { return m_maxEnemyCount; }
 
 	private:
 		void FixWaveNum();
@@ -68,11 +68,11 @@ namespace ButiEngine {
 		bool m_isSceneStart;
 
 		//ステージクリアまでの敵討伐カウント
-		std::int16_t m_enemyDeadCount;
+		std::int32_t m_enemyDeadCount;
 		//ステージクリアまでの敵討伐カウント最大値
-		std::int16_t m_maxEnemyCount;
+		std::int32_t m_maxEnemyCount;
 		//〇体まで出現させるためのカウント
-		std::int16_t m_enemySpawnCount;
+		std::int32_t m_enemySpawnCount;
 	};
 
 }
